requirements.txt
opencv-python>=4.12.0
mediapipe>=0.10.14
numpy>=2.2.6
Pillow>=10.0.0

